Refactor (#426)
Currently Local Node is quite a mess, it's hard to maintain, add new features. In the past it was responsible only for starting couple of docker containers and that's it. Now apart from that, it needs to do this in different kind of modes, while activate/deactivate features, as well as generating accounts and more.

This PR aims to refactor the whole architecture by making the code easy to understand, maintain and extend.

By utilizing state machine pattern, we introduce most importantly separation of concerns, which makes sure that each and every state does only what it's supposed to do and only goes ahead to the next if it's job is completed.
And service locator pattern, allows us to easly create services which are always available due to the nature of the service locator(being singleton) and easy to reach. That way we don't have to increase coupling or duplicate code to use different kind of services.

Quick explanation:
For each configuration(start,stop,restart,generate-accounts) there is a predefined set of states (`StateData`), which are being loaded in the `StateController` upon startup. By cycling through the states, each and every one do only what it's supposed to and emit event when it's ready (event can be either finish for success or error). Most of the logic for each state is copied from the old code, it's just seprated logically. 

`Bootstraper` is used to create the Service Locator and register all services. After this all we need to do, to access a service is to call it by using the Service Locator singleton.
Current Services:

- `CLIService` is used to load all cli options and make them easy to access, whethever we need them
- `ClientService` is used to create a SDK client, which is accessable from anywhere.
- `LoggerService` is used to log all msg to the console, supports verbose and will be used together with the TUI
- `ConnectionService` have helper methods which are responsible for connection checks
- `DockerService` have all kind of docker related utils

Things that does not work on previous version and we can delay fixing, without delaying release:
- Multi node
- Debug

To start the node we can use `hedera` keyword (when it's released) or npm run `start`/`stop`/`restart`/`generate-accounts`
If we want to add options together with our start command we need to first add `--`. 
For example, if we want to start the node in full mode, we'd write `npm run start -- --full`

Signed-off-by: georgi-l95 <glazarov95@gmail.com>